v0.3
- Major refactor
- cache has its own module
- brew improvements (this will be replaced by hteacup)
