+++ config.toml
@@ -4,8 +4,9 @@ host = "0.0.0.0"
 threads = 4
 root = "public"
 cache = true
-cache_ttl = 3600
+cache_ttl = 36
 [proxy]
 "/test" = "http://example.com"
 "/google" = "http://google.com"
+"/myip" = "http://ifconfig.co"
 # "/" = "http://ifconfig.co" # this will override all the proxys and local request
