+++ public/index.html
@@ -1,16 +1,19 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
-<head>
-  <title>HTeaPot Server</title>
-  <script src="index.js"></script>
-</head>
-<body>
-  <h1>HTeaPot Server</h1>
-  <p>Congratulations! HTeaPot, the HTTP server, is correctly installed and configured.</p>
-  <p>You are now ready to serve your web applications.</p>
-  <button onclick="brew()" >Brew!</button>
-  <br>
-  <br>
-  <img src="favicon.ico" alt="tea"/>
-</body>
+    <head>
+        <title>HTeaPot Server</title>
+        <script src="index.js"></script>
+    </head>
+    <body>
+        <h1>HTeaPot Server</h1>
+        <p>
+            Congratulations! HTeaPot, the HTTP server, is correctly installed
+            and configured.
+        </p>
+        <p>You are now ready to serve your web applications.</p>
+        <button onclick="brew()">Brew!</button>
+        <br />
+        <br />
+        <img src="favicon.ico" alt="tea" />
+    </body>
 </html>
