+++ src/main.rs
@@ -1,24 +1,88 @@
 mod brew;
+mod cache;
 mod config;
 pub mod hteapot;
 mod logger;
 
-use std::collections::HashMap;
 use std::fs;
 use std::io;
+use std::path::Path;
 use std::sync::Mutex;
-use std::time;
-use std::time::SystemTime;
 
 use brew::fetch;
+use cache::Cache;
+use config::Config;
 use hteapot::{Hteapot, HttpResponse, HttpStatus};
 
 use logger::Logger;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
+fn is_proxy(config: &Config, path: String) -> Option<String> {
+    for proxy_path in config.proxy_rules.keys() {
+        let path_proxy = path.strip_prefix(proxy_path);
+        if path_proxy.is_some() {
+            let path_proxy = path_proxy.unwrap();
+            let url = config.proxy_rules.get(proxy_path).unwrap();
+            let separator = if path_proxy.starts_with('/') || url.ends_with('/') {
+                ""
+            } else {
+                "/"
+            };
+            let url = format!("{}{}{}", url, separator, path_proxy);
+            return Some(url);
+        }
+    }
+    None
+}
+
+fn serve_proxy(proxy_url: String) -> HttpResponse {
+    let raw_response = fetch(&proxy_url);
+    match raw_response {
+        Ok(raw) => HttpResponse::new_raw(raw),
+        Err(_) => HttpResponse::new(HttpStatus::NotFound, "not found", None),
+    }
+}
+
+fn get_mime_tipe(config: &Config, path: String) -> String {
+    let mut path = format!("{}{}", config.root, path);
+    if Path::new(path.as_str()).is_dir() {
+        let separator = if path.ends_with('/') { "" } else { "/" };
+        path = format!("{}{}{}", path, separator, config.index);
+    }
+    let extension = Path::new(path.as_str())
+        .extension()
+        .unwrap()
+        .to_str()
+        .unwrap();
+    let mimetipe = match extension {
+        "js" => "text/javascript",
+        "json" => "application/json",
+        "css" => "text/css",
+        "html" => "text/html",
+        _ => "text/plain",
+    };
+
+    mimetipe.to_string()
+}
+
+fn serve_file(config: &Config, path: String) -> Option<Vec<u8>> {
+    let mut path = format!("{}{}", config.root, path);
+    if Path::new(path.as_str()).is_dir() {
+        let separator = if path.ends_with('/') { "" } else { "/" };
+        path = format!("{}{}{}", path, separator, config.index);
+    }
+    let r = fs::read(path);
+    if r.is_ok() {
+        Some(r.unwrap())
+    } else {
+        None
+    }
+}
+
 fn main() {
     let args = std::env::args().collect::<Vec<String>>();
+    let mut serving_path = None;
     if args.len() >= 2 {
         match args[1].as_str() {
             "--help" | "-h" => {
@@ -30,18 +94,45 @@ fn main() {
                 println!("Hteapot {}", VERSION);
                 return;
             }
-            _ => "",
+            "--serve" | "-s" => {
+                serving_path = Some(args.get(2).unwrap().clone());
+            }
+            _ => (),
         };
     }
 
-    let config = if args.len() > 1 {
+    let config = if args.len() == 2 {
         config::Config::load_config(&args[1])
+    } else if serving_path.is_some() {
+        let serving_path_str = serving_path.unwrap();
+        let serving_path_str = serving_path_str.as_str();
+        let serving_path = Path::new(serving_path_str);
+        let mut c = config::Config::new_default();
+        c.host = "0.0.0.0".to_string();
+        if serving_path.is_dir() {
+            c.root = serving_path.to_str().unwrap_or_default().to_string();
+        } else {
+            c.index = serving_path
+                .file_name()
+                .unwrap()
+                .to_str()
+                .unwrap_or_default()
+                .to_string();
+            c.root = serving_path
+                .parent()
+                .unwrap_or(Path::new("./"))
+                .to_str()
+                .unwrap_or_default()
+                .to_string();
+        }
+        c
     } else {
         config::Config::new_default()
     };
+
     let proxy_only = config.proxy_rules.get("/").is_some();
     let logger = Mutex::new(Logger::new(io::stdout()));
-    let cache: Mutex<HashMap<String, (Vec<u8>, u64)>> = Mutex::new(HashMap::new());
+    let cache: Mutex<Cache> = Mutex::new(Cache::new(config.cache_ttl as u64));
     let server = Hteapot::new_threaded(config.host.as_str(), config.port, config.threads);
     logger.lock().expect("this doesnt work :C").msg(format!(
         "Server started at http://{}:{}",
@@ -61,117 +152,40 @@ fn main() {
     }
 
     server.listen(move |req| {
+        // SERVER CORE
+        // for each request
+
         logger.lock().expect("this doesnt work :C").msg(format!(
             "Request {} {}",
             req.method.to_str(),
             req.path
         ));
-        let path = if req.path.ends_with("/") {
-            let mut path = req.path.clone();
-            path.push_str(&config.index);
-            path
-        } else {
-            req.path.clone()
-        };
-        let path_clone = req.path.clone();
-        let divided_path: Vec<&str> = path_clone.split('/').skip(1).collect();
-        if divided_path.is_empty() {
-            return HttpResponse::new(HttpStatus::BadRequest, b"Invalid path", None);
-        }
 
-        let first_one = format!("/{}", divided_path[0]);
-        let rest_path = divided_path[1..].join("/");
+        let is_proxy = is_proxy(&config, req.path.clone());
 
-        if proxy_only || config.proxy_rules.contains_key(&first_one) {
-            let url = if proxy_only {
-                let url = config.proxy_rules.get("/").unwrap();
-                if rest_path.len() != 0 {
-                    format!("{}{}/{}", url, first_one, rest_path)
-                } else {
-                    format!("{}{}", url, first_one)
-                }
-            } else {
-                let url = config.proxy_rules.get(&first_one).unwrap();
-                format!("{}/{}", url, rest_path)
-            };
-            logger
-                .lock()
-                .expect("this doesnt work :C")
-                .msg(format!("Proxying to: {}", url));
-            return match fetch(&url) {
-                Ok(response) => HttpResponse::new_raw(response),
-                Err(err) => {
-                    HttpResponse::new(HttpStatus::InternalServerError, err.as_bytes(), None)
-                }
-            };
+        if proxy_only || is_proxy.is_some() {
+            return serve_proxy(is_proxy.unwrap());
         }
-        let path = format!("./{}/{}", config.root, path);
-        let cache_result = {
-            if config.cache {
-                let cache = cache.lock();
-                if cache.is_err() {
-                    None
-                } else {
-                    let cache = cache.unwrap();
-                    let r = cache.get(&path);
-                    match r {
-                        Some(r) => Some(r.clone()),
-                        None => None,
-                    }
+        if !Path::new(req.path.clone().as_str()).exists() {
+            return HttpResponse::new(HttpStatus::NotFound, "Not found", None);
+        }
+        let mimetype = get_mime_tipe(&config, req.path.clone());
+        let content: Option<Vec<u8>> = if config.cache {
+            let mut cachee = cache.lock().expect("Error locking cache");
+            let mut r = cachee.get(req.path.clone());
+            if r.is_none() {
+                r = serve_file(&config, req.path.clone());
+                if r.is_some() {
+                    cachee.set(req.path.clone(), r.clone().unwrap());
                 }
-            } else {
-                None
             }
-        };
-        let mut is_cache = false;
-        let content: Result<Vec<u8>, _> = if cache_result.is_some() {
-            let (content, ttl) = cache_result.unwrap();
-            let now = SystemTime::now();
-            let since_epoch = now
-                .duration_since(time::UNIX_EPOCH)
-                .expect("Time went backwards");
-            let secs = since_epoch.as_secs();
-            if secs > ttl {
-                fs::read(&path)
-            } else {
-                is_cache = true;
-                Ok(content)
-            }
+            r
         } else {
-            fs::read(&path)
+            serve_file(&config, req.path)
         };
         match content {
-            Ok(content) => {
-                if config.cache {
-                    let cache = cache.lock();
-                    if cache.is_ok() && is_cache {
-                        let mut cache = cache.unwrap();
-                        let now = SystemTime::now();
-                        let since_epoch = now
-                            .duration_since(time::UNIX_EPOCH)
-                            .expect("Time went backwards");
-                        let secs = since_epoch.as_secs() + config.cache_ttl as u64;
-                        cache.insert(path, (content.clone(), secs));
-                    }
-                }
-                return HttpResponse::new(HttpStatus::OK, &content, None);
-            }
-            Err(e) => match e.kind() {
-                io::ErrorKind::NotFound => {
-                    return HttpResponse::new(
-                        HttpStatus::NotFound,
-                        "<h1> 404 Not Found </h1>",
-                        headers!("Content-Type" => "text/html", "Server" => "HteaPot"),
-                    );
-                }
-                _ => {
-                    return HttpResponse::new(
-                        HttpStatus::InternalServerError,
-                        "<h1> 500 Internal Server Error </h1>",
-                        headers!("Content-Type" => "text/html", "Server" => "HteaPot"),
-                    );
-                }
-            },
+            Some(c) => HttpResponse::new(HttpStatus::OK, c, headers!("Content-Type" => mimetype)),
+            None => HttpResponse::new(HttpStatus::NotFound, "Not found", None),
         }
     });
 }
